pyproject.toml
[project]
name = "streamlittrainingapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-firestore>=2.21.0",
    "streamlit[auth]>=1.51.0",
    "streamlit-browser-engine>=0.0.3",
    "firebase-admin>=7.1.0",
]
